+++ frontend/src/pages/admin.css
@@ -1,0 +1,11 @@
+.admin-link {
+    color: #fff;
+    text-decoration: none;
+    font-size: 1.2em;
+    font-weight: bold;
+}
+
+.admin-link:hover {
+    color: #fff;
+    text-decoration: underline;
+}
