+++ frontend/src/pages/admin/Licenses.js
@@ -1,0 +1,14 @@
+import React, { useEffect } from "react";
+
+export const AdminLicenses = () => {
+
+  useEffect(() => {
+    document.title = "Bürgerbüro - Genehmigungen";
+  }, []);
+
+  return (
+    <div>
+      <h1>Genehmigungen</h1>
+    </div>
+  );
+};
